+++ Runtime/LevelScripting/Logic/PlatformLogic.cs
@@ -3,6 +3,7 @@ using System.Collections;
 using System.Linq;
 using UnityEngine;
 using UnityEngine.Events;
+using UnityEngine.Serialization;
 
 namespace GameplayIngredients.Logic
 {
@@ -19,14 +20,20 @@ namespace GameplayIngredients.Logic
 
         public InclusionMode inclusionMode = InclusionMode.IsTarget;
 
+        [ReorderableList, FormerlySerializedAs("Calls")]
+        public Callable[] OnTestValid;
+
         [ReorderableList]
-        public Callable[] Calls;
+        public Callable[] OnTestInvalid;
+
 
         public override void Execute(GameObject instigator = null)
         {
 
             if(platforms.Contains(Application.platform) == (inclusionMode == InclusionMode.IsTarget))
-                Call(Calls, instigator);
+                Call(OnTestValid, instigator);
+            else
+                Call(OnTestInvalid, instigator);
         }
     }
 }