+++ Editor/SceneViewToolbar.cs
@@ -101,12 +101,18 @@ namespace GameplayIngredients.Editor
                         SceneViewPOV.ShowPopup(btnrect, sceneView);
                     }
 
-                    GUILayout.FlexibleSpace();
+                    // Check Window
+                    GUILayout.Space(16);
+                    if (GUILayout.Button(Contents.checkWindow, EditorStyles.toolbarButton))
+                    {
+                        EditorWindow.GetWindow<CheckWindow>();
+                    }
 
                     // Custom Code here
                     if (OnSceneViewToolbarGUI != null)
                         OnSceneViewToolbarGUI.Invoke(sceneView);
 
+                    GUILayout.FlexibleSpace();
                     // Saving Space not to overlap view controls
                     GUILayout.Space(96);
 
@@ -150,6 +156,7 @@ namespace GameplayIngredients.Editor
             public static GUIContent lockLinkGameView;
             public static GUIContent linkGameView;
             public static GUIContent linkGameViewCinemachine;
+            public static GUIContent checkWindow;
 
             static Contents()
             {
@@ -160,10 +167,11 @@ namespace GameplayIngredients.Editor
                 linkGameViewCinemachine = new GUIContent(EditorGUIUtility.Load("Packages/net.peeweek.gameplay-ingredients/Icons/GUI/Camera16x16.png") as Texture);
                 linkGameViewCinemachine.text = " Cine";
 
-
-
                 playFromHere = new GUIContent(EditorGUIUtility.IconContent("Animation.Play"));
                 playFromHere.text = "Here";
+
+                checkWindow = new GUIContent(EditorGUIUtility.IconContent("Valid"));
+                checkWindow.text = "Check";
             }
         }