+++ CHANGELOG.md
@@ -5,10 +5,14 @@
 #### Added
 
 - Added API Access to the Link Game View Camera object
+- Added Valid/Invalid paths for Platform Logic
+- Added Local Space control for ReachPositionRig
+- Added "Update SetStateAction" Button in State Machine components to populate the game object with Set State Actions
 
 #### Fixed
 
 * Fixed Possible Nulls in Discover Asset reference
+* Fixed Possible Out-of range director time in DirectorControlRig
 
 ## 2019.3.5