+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added Valid/Invalid paths for Platform Logic
 - Added Local Space control for ReachPositionRig
 - Added "Update SetStateAction" Button in State Machine components to populate the game object with Set State Actions
+- Added **Check/Resolve** Window and API : The check window enables performing scriptable checks and resolution actions in your scene for maintenance purposes.
 
 #### Fixed